requirements.txt
discord.py>=2.0.0
aiohttp>=3.10.11
python-dotenv>=0.20.0
Pillow>=9.0.0
